# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form operations
  - name: projects
    description: Public project read endpoints
  - name: admin-projects
    description: Authenticated project administration endpoints
  - name: storage
    description: Object storage upload and serving endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      description: Sends a contact form submission to the firm via email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRequest"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List published projects
      description: Returns all projects ordered by sortOrder, including each project's gallery images.
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  /projects/{slug}:
    get:
      operationId: getProjectBySlug
      tags: [projects]
      summary: Get project by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/projects:
    get:
      operationId: adminListProjects
      tags: [admin-projects]
      summary: List all projects (admin)
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: adminCreateProject
      tags: [admin-projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/projects/{id}:
    get:
      operationId: adminGetProject
      tags: [admin-projects]
      summary: Get project by id (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: adminUpdateProject
      tags: [admin-projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: adminDeleteProject
      tags: [admin-projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/projects/{id}/images:
    put:
      operationId: adminReplaceProjectImages
      tags: [admin-projects]
      summary: Replace gallery images for a project
      description: Replaces the entire ordered gallery for the project with the provided list.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceImagesRequest"
      responses:
        "200":
          description: Updated project with new gallery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to generate upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /storage/public-objects/{filePath}:
    get:
      tags: [storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed with correct Content-Type
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve an uploaded object entity
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed with correct Content-Type
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ContactRequest:
      type: object
      properties:
        nombre:
          type: string
          minLength: 2
          maxLength: 120
          description: Nombre completo del remitente
        email:
          type: string
          format: email
          maxLength: 200
        telefono:
          type: string
          maxLength: 40
          description: Teléfono opcional
        mensaje:
          type: string
          minLength: 10
          maxLength: 4000
      required:
        - nombre
        - email
        - mensaje
    ContactResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ProjectImage:
      type: object
      required: [id, imagePath, sortOrder]
      properties:
        id:
          type: string
          format: uuid
        imagePath:
          type: string
          description: Either an absolute public path (e.g. /images/foo.jpg) or an object storage path (e.g. /objects/uploads/uuid).
        sortOrder:
          type: integer

    Project:
      type: object
      required:
        - id
        - slug
        - title
        - year
        - location
        - type
        - description
        - coverImagePath
        - sortOrder
        - images
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        year:
          type: integer
        location:
          type: string
          description: Project location category (e.g. quinta-region, region-metropolitana, argentina).
        type:
          type: string
          description: Project type category (e.g. residencial-casa, residencial-departamento, oficinas).
        description:
          type: string
        coverImagePath:
          type: string
        sortOrder:
          type: integer
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProjectImage"

    ProjectInput:
      type: object
      required:
        - slug
        - title
        - year
        - location
        - type
        - coverImagePath
      properties:
        slug:
          type: string
          minLength: 1
          maxLength: 120
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        title:
          type: string
          minLength: 1
          maxLength: 200
        year:
          type: integer
          minimum: 1900
          maximum: 2100
        location:
          type: string
          enum: [quinta-region, region-metropolitana, argentina]
        type:
          type: string
          enum: [residencial-casa, residencial-departamento, oficinas]
        description:
          type: string
          maxLength: 4000
        coverImagePath:
          type: string
          minLength: 1
        sortOrder:
          type: integer

    ReplaceImagesRequest:
      type: object
      required: [images]
      properties:
        images:
          type: array
          items:
            type: object
            required: [imagePath]
            properties:
              imagePath:
                type: string
                minLength: 1

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"