# lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectImage } from "./projectImage";

export interface Project {
  id: string;
  slug: string;
  title: string;
  year: number;
  /** Project location category (e.g. quinta-region, region-metropolitana, argentina). */
  location: string;
  /** Project type category (e.g. residencial-casa, residencial-departamento, oficinas). */
  type: string;
  description: string;
  coverImagePath: string;
  sortOrder: number;
  images: ProjectImage[];
}