# lib/api-zod/src/generated/types/projectImage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectImage {
  id: string;
  /** Either an absolute public path (e.g. /images/foo.jpg) or an object storage path (e.g. /objects/uploads/uuid). */
  imagePath: string;
  sortOrder: number;
}