# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectInputLocation } from "./projectInputLocation";
import type { ProjectInputType } from "./projectInputType";

export interface ProjectInput {
  /**
   * @minLength 1
   * @maxLength 120
   * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
   */
  slug: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  title: string;
  /**
   * @minimum 1900
   * @maximum 2100
   */
  year: number;
  location: ProjectInputLocation;
  type: ProjectInputType;
  /** @maxLength 4000 */
  description?: string;
  /** @minLength 1 */
  coverImagePath: string;
  sortOrder?: number;
}